+++ App.tsx
@@ -1,4 +1,4 @@
-import Home from "./src/pages/home";
+import Home from "./src/pages/home/home";
 // import Login from "./src/pages/login";
 
 export default function App() {
